InventoryItem ADMIN Functions

+++ src/Controller/Main/InventoryItemController.php
@@ -6,12 +6,13 @@ use App\Controller\CustomAbstractController;
 use App\Entity\InventoryItem;
 use App\Form\InventoryItemType;
 use App\Repository\InventoryItemRepository;
-use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use App\Repository\InventoryRepository;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use Symfony\Component\Routing\Annotation\Route;
 
-#[Route('/inventory/item')]
+#[Route('/inventory')]
 class InventoryItemController extends CustomAbstractController
 {
     #[Route(
@@ -32,7 +33,7 @@ class InventoryItemController extends CustomAbstractController
         ]);
     }
 
-    #[Route('/new',
+    #[Route('/{id}/new',
         name: 'app_inventory_item_new',
         options: [
             "system" => 'false'
@@ -42,15 +43,20 @@ class InventoryItemController extends CustomAbstractController
             "role" => "superadmin"
         ],
         methods: ['GET', 'POST'])]
-    public function new(Request $request, InventoryItemRepository $inventoryItemRepository): Response
+    public function new(Request $request, int $id, InventoryRepository $inventoryRepository, InventoryItemRepository $inventoryItemRepository): Response
     {
         $inventoryItem = new InventoryItem();
+        $inventory = $inventoryRepository->find($id);
+        if (empty($inventory))
+            throw new NotFoundHttpException('Inventory ' . $id . ' not found');
+        $inventoryItem->setInventory($inventory);
         $form = $this->createForm(InventoryItemType::class, $inventoryItem);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $inventoryItemRepository->add($inventoryItem);
-            return $this->redirectToRoute('app_inventory_item_index', [], Response::HTTP_SEE_OTHER);
+            $this->addFlash('success', 'Inventory Item is created for inventory #' . $id);
+            return $this->redirectToRoute('app_inventory_show', ['id' => $id], Response::HTTP_SEE_OTHER);
         }
 
         return $this->renderForm('inventory_item/new.html.twig', [
@@ -60,25 +66,7 @@ class InventoryItemController extends CustomAbstractController
     }
 
     #[Route(
-        '/{id}',
-        name: 'app_inventory_item_show',
-        options: [
-            "system" => 'false'
-        ],
-        defaults: [
-            "description" => "Detailed info for a given inventory items",
-            "role" => "superadmin"
-        ],
-        methods: ['GET'])]
-    public function show(InventoryItem $inventoryItem): Response
-    {
-        return $this->render('inventory_item/show.html.twig', [
-            'inventory_item' => $inventoryItem,
-        ]);
-    }
-
-    #[Route(
-        '/{id}/edit',
+        '/{inventoryId}/{inventoryItemId}/edit',
         name: 'app_inventory_item_edit',
         options: [
             "system" => 'false'
@@ -88,24 +76,29 @@ class InventoryItemController extends CustomAbstractController
             "role" => "superadmin"
         ],
         methods: ['GET', 'POST'])]
-    public function edit(Request $request, InventoryItem $inventoryItem, InventoryItemRepository $inventoryItemRepository): Response
+    public function edit(Request $request, int $inventoryId, int $inventoryItemId, InventoryItemRepository $inventoryItemRepository): Response
     {
+        $inventoryItem = $inventoryItemRepository->find($inventoryItemId);
+        if (empty($inventoryItem))
+            throw new NotFoundHttpException('Inventory Item #' . $inventoryItemId . ' Not Found');
         $form = $this->createForm(InventoryItemType::class, $inventoryItem);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $inventoryItemRepository->add($inventoryItem);
-            return $this->redirectToRoute('app_inventory_item_index', [], Response::HTTP_SEE_OTHER);
+            $this->addFlash('success', 'Item #' . $inventoryItemId . ' updated');
+            return $this->redirectToRoute('app_inventory_show', ['id' => $inventoryId], Response::HTTP_SEE_OTHER);
         }
 
         return $this->renderForm('inventory_item/edit.html.twig', [
             'inventory_item' => $inventoryItem,
+            'inventory_id' => $inventoryId,
             'form' => $form,
         ]);
     }
 
     #[Route(
-        '/{id}',
+        '/{inventoryId}/{inventoryItemId}/delete',
         name: 'app_inventory_item_delete',
         options: [
             "system" => 'false'
@@ -116,12 +109,16 @@ class InventoryItemController extends CustomAbstractController
         ],
         methods: ['POST'])
     ]
-    public function delete(Request $request, InventoryItem $inventoryItem, InventoryItemRepository $inventoryItemRepository): Response
+    public function delete(Request $request, int $inventoryId, int $inventoryItemId, InventoryItemRepository $inventoryItemRepository): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$inventoryItem->getId(), $request->request->get('_token'))) {
+        $inventoryItem = $inventoryItemRepository->find($inventoryItemId);
+        if (empty($inventoryItem))
+            throw new  NotFoundHttpException('Inventory Item #' . $inventoryItemId . ' Not Found');
+        if ($this->isCsrfTokenValid('delete' . $inventoryItem->getId(), $request->request->get('_token'))) {
             $inventoryItemRepository->remove($inventoryItem);
+            $this->addFlash('success', 'Item #' . $inventoryItemId . ' removed');
         }
 
-        return $this->redirectToRoute('app_inventory_item_index', [], Response::HTTP_SEE_OTHER);
+        return $this->redirectToRoute('app_inventory_show', ['id' => $inventoryId], Response::HTTP_SEE_OTHER);
     }
 }
